requirements.txt
djangorestframework==3.14.0
django==4.1
python-dotenv==1.0.0
mysqlclient==2.2.1
channels==4.0.0
channels-redis==4.1.0